fix: add static params for product pages
